+++ src/fde.cpp
@@ -14,21 +14,21 @@
 #include "ann_exception.h"
 #include "utils.h"
 
-double dot_product(const std::vector<double>& h, const std::vector<double>& p, size_t dimensions) {
+float dot_product(const std::vector<float>& h, const std::vector<float>& p, size_t dimensions) {
     // REQUIRES: h.size() == dimensions && p.size() == dimensions
-    double result = 0.0;
+    float result = 0.0;
     for (size_t i = 0; i < dimensions; i++)
         result += h[i] * p[i];
     return result;
 };
 
 // TODO: Implement SIMD optimizations
-// TODO: Use type template to support datatypes other than double floats.
-std::vector<double> SimHash::generate_gaussian_vector(size_t d) {
+// TODO: Use type template to support datatypes other than float floats.
+std::vector<float> SimHash::generate_gaussian_vector(size_t d) {
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::normal_distribution<double>dist(0.0, 1.0); // standard Gaussian
-    std::vector<double> result(d);
+    std::normal_distribution<float>dist(0.0, 1.0); // standard Gaussian
+    std::vector<float> result(d);
     for (size_t i = 0; i < d; i++)
         result[i] = dist(gen);
     return result;
@@ -41,8 +41,8 @@ SimHash::SimHash(size_t dimensions, size_t k_sim): AbstractLSH(dimensions, k_sim
     }
 };
 
-uint64_t SimHash::compute_hash(const std::vector<double>& v) const {
-    uint64_t hash = 0;
+uint32_t SimHash::compute_hash(const std::vector<float>& v) const {
+    uint32_t hash = 0;
     for (size_t i = 0; i < k_sim; i++) {
         if (dot_product(hyperplanes[i], v, dimensions) >= 0) {
             hash |= (1ULL << i); // Little Endian
@@ -55,16 +55,16 @@ ExactChamferSimilarity::ExactChamferSimilarity(size_t dimensions): AbstractChamf
 
 // TODO: SIMD optimizations
 // TODO: Make sure that embeddings are normalized
-double ExactChamferSimilarity::compute_similarity(
-    const std::vector<std::vector<double>>& P,
-    const std::vector<std::vector<double>>& Q) const {
-    double result = 0.0;
+float ExactChamferSimilarity::compute_similarity(
+    const std::vector<std::vector<float>>& P,
+    const std::vector<std::vector<float>>& Q) const {
+    float result = 0.0;
     // TODO: [fine grained performance engineering] :change the ordering of
     //       this iteration based on the relative size of P and Q
     for (auto p: P) {
-        double best = 0.0;
+        float best = 0.0;
         for (auto q: Q) {
-            double c = dot_product(p, q, dimensions);
+            float c = dot_product(p, q, dimensions);
             best = c > best ? c : best;
         }
         result += best;
@@ -73,13 +73,13 @@ double ExactChamferSimilarity::compute_similarity(
 };
 
 
-std::vector<std::vector<double>> FDESimilarity::get_scaled_S() { // (1 / sqrt(d_proj))S
+std::vector<std::vector<float>> FDESimilarity::get_scaled_S() { // (1 / sqrt(d_proj))S
     std::random_device rd;
     std::mt19937 gen(rd());
     std::uniform_int_distribution<int> binary_dist(0, 1);
 
-    double scale = 1.0 / std::sqrt(d_proj);
-    std::vector<std::vector<double>> result(d_proj, std::vector<double>(dimensions));
+    float scale = 1.0 / std::sqrt(d_proj);
+    std::vector<std::vector<float>> result(d_proj, std::vector<float>(dimensions));
 
     for (int i = 0; i < d_proj; ++i) {
         for (int j = 0; j < dimensions; ++j) {
@@ -90,17 +90,17 @@ std::vector<std::vector<double>> FDESimilarity::get_scaled_S() { // (1 / sqrt(d_
     return result;
 };
 
-uint64_t FDESimilarity::compute_hash_from_rep_idx(size_t idx, const std::vector<double>& v) const {
+uint32_t FDESimilarity::compute_hash_from_rep_idx(size_t idx, const std::vector<float>& v) const {
     // REQUIRES: 0 <= idx < r_reps && v.size() == dimensions
     return all_simhash[idx].compute_hash(v);
 };
 
-std::vector<double> FDESimilarity::compute_proj_from_rep_idx(size_t idx, const std::vector<double> &v) const {
+std::vector<float> FDESimilarity::compute_proj_from_rep_idx(size_t idx, const std::vector<float> &v) const {
     // TODO: Perf engineer this
     // REQUIRES: d_proj == all_S[idx].size();
     // REQUIRES: dimensions == v.size() == all_S[0][0].size()
 
-    std::vector<double> result(d_proj, 0.0);
+    std::vector<float> result(d_proj, 0.0);
     for (size_t i = 0; i < d_proj; i++) {
         result[i] = dot_product(all_S[idx][i], v, dimensions);
     }
@@ -108,46 +108,46 @@ std::vector<double> FDESimilarity::compute_proj_from_rep_idx(size_t idx, const s
 };
 
 
-std::vector<double> FDESimilarity::encode_document_once(size_t idx, const std::vector<std::vector<double>> &P) const {
+std::vector<float> FDESimilarity::encode_document_once(size_t idx, const std::vector<std::vector<float>> &P) const {
     // idx is the repetition index
-    std::vector<std::vector<double>> P_hash_grouped;
+    std::vector<std::vector<float>> P_hash_grouped;
     std::vector<size_t> bucket_counts(B, 0);
     P_hash_grouped.resize(B);
     for (size_t i = 0; i < B; i++)
-        P_hash_grouped[i] = std::vector<double>(dimensions, 0.0);
+        P_hash_grouped[i] = std::vector<float>(dimensions, 0.0);
     for (auto p: P) {
-        uint64_t hash_value = compute_hash_from_rep_idx(idx, p);
+        uint32_t hash_value = compute_hash_from_rep_idx(idx, p);
         bucket_counts[hash_value] ++;
-        // TODO: Double-check the type conversion here
+        // TODO: float-check the type conversion here
         for (size_t j = 0; j < dimensions; j++) {
             P_hash_grouped[hash_value][j] = P_hash_grouped[hash_value][j] * (bucket_counts[hash_value] - 1 ) + p[j];
             P_hash_grouped[hash_value][j] /= bucket_counts[hash_value];
         }
     }
-    std::vector<double> P_phi;
+    std::vector<float> P_phi;
     P_phi.reserve(B * d_proj);
     for (size_t i = 0; i < B; i++) {
-        std::vector<double> projection = compute_proj_from_rep_idx(idx, P_hash_grouped[i]);
+        std::vector<float> projection = compute_proj_from_rep_idx(idx, P_hash_grouped[i]);
         P_phi.insert(P_phi.end(), projection.begin(), projection.end());
     }
     return P_phi;
 };
 
-std::vector<double> FDESimilarity::encode_query_once(size_t idx, const std::vector<std::vector<double>> &Q) const {
+std::vector<float> FDESimilarity::encode_query_once(size_t idx, const std::vector<std::vector<float>> &Q) const {
     // idx is the repetition index
-    std::vector<std::vector<double>> Q_hash_grouped;
+    std::vector<std::vector<float>> Q_hash_grouped;
     Q_hash_grouped.resize(B);
     for (size_t i = 0; i < B; i++)
-        Q_hash_grouped[i] = std::vector<double>(dimensions, 0.0);
+        Q_hash_grouped[i] = std::vector<float>(dimensions, 0.0);
     for (auto q: Q) {
-        uint64_t hash_value = compute_hash_from_rep_idx(idx, q);
-        // TODO: Double-check the type conversion here
+        uint32_t hash_value = compute_hash_from_rep_idx(idx, q);
+        // TODO: float-check the type conversion here
         for (size_t j = 0; j < dimensions; j++) Q_hash_grouped[hash_value][j] += q[j];
     }
-    std::vector<double> Q_phi;
+    std::vector<float> Q_phi;
     Q_phi.reserve(B * d_proj);
     for (size_t i = 0; i < B; i++) {
-        std::vector<double> projection = compute_proj_from_rep_idx(idx, Q_hash_grouped[i]);
+        std::vector<float> projection = compute_proj_from_rep_idx(idx, Q_hash_grouped[i]);
         Q_phi.insert(Q_phi.end(), projection.begin(), projection.end());
     }
     return Q_phi;
@@ -157,22 +157,22 @@ size_t FDESimilarity::get_d_fde() {
     return B * d_proj * r_reps;
 };
 
-std::vector<double> FDESimilarity::encode_document(const std::vector<std::vector<double>> &P) const {
+std::vector<float> FDESimilarity::encode_document(const std::vector<std::vector<float>> &P) const {
     // TODO: implement fill_empty_clusters
-    std::vector<double> result;
+    std::vector<float> result;
     result.reserve(B * d_proj * r_reps);
     for(size_t idx = 0; idx < r_reps; idx++) {
-        std::vector<double> trial = encode_query_once(idx, P);
+        std::vector<float> trial = encode_query_once(idx, P);
         result.insert(result.end(), trial.begin(), trial.end());
     }
     return result;
 };
 
-std::vector<double> FDESimilarity::encode_query(const std::vector<std::vector<double>> &Q) const {
-    std::vector<double> result;
+std::vector<float> FDESimilarity::encode_query(const std::vector<std::vector<float>> &Q) const {
+    std::vector<float> result;
     result.reserve(B * d_proj * r_reps);
     for(size_t idx = 0; idx < r_reps; idx++) {
-        std::vector<double> trial = encode_query_once(idx, Q);
+        std::vector<float> trial = encode_query_once(idx, Q);
         result.insert(result.end(), trial.begin(), trial.end());
     }
     return result;
@@ -191,9 +191,9 @@ k_sim(_k_sim), r_reps(_r_reps) {
     }
 };
 
-double FDESimilarity::compute_similarity(
+float FDESimilarity::compute_similarity(
     // TODO: implement final projections
-    const std::vector<std::vector<double>>& P,
-    const std::vector<std::vector<double>>& Q) const {
+    const std::vector<std::vector<float>>& P,
+    const std::vector<std::vector<float>>& Q) const {
     return dot_product(encode_document(P), encode_query(Q), B * d_proj * r_reps);
 };
